# SourceCode/front-end/logsmart/openapi.json
{
	"openapi": "3.1.0",
	"info": { "title": "back-end", "description": "", "license": { "name": "" }, "version": "1.1.0" },
	"paths": {
		"/auth/admin/remove-member": {
			"delete": {
				"tags": ["Company Management"],
				"summary": "Deletes a company member (admin only).",
				"description": "# Errors\nReturns an error if the user is not an admin or if the deletion fails.",
				"operationId": "admin_delete_member",
				"requestBody": {
					"content": {
						"application/json": { "schema": { "$ref": "#/components/schemas/RemoveMemberRequest" } }
					},
					"required": true
				},
				"responses": {
					"200": { "description": "Member deleted successfully" },
					"400": {
						"description": "Invalid request",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					},
					"403": {
						"description": "Forbidden - not an admin or different company",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					},
					"404": {
						"description": "User not found",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					}
				},
				"security": [{ "bearer_auth": [] }]
			}
		},
		"/auth/admin/update-member": {
			"put": {
				"tags": ["Company Management"],
				"summary": "Updates a company member's profile (admin only).",
				"description": "# Errors\nReturns an error if the user is not an admin, the request is invalid, or the update fails.",
				"operationId": "admin_update_member_profile",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": { "$ref": "#/components/schemas/AdminUpdateMemberRequest" }
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "Member profile updated successfully",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/GetCompanyMembersResponse" }
							}
						}
					},
					"400": {
						"description": "Invalid request",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					},
					"403": {
						"description": "Forbidden - not an admin or different company",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					},
					"404": {
						"description": "User not found",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					}
				},
				"security": [{ "bearer_auth": [] }]
			}
		},
...
		"/auth/company/members": {
			"get": {
				"tags": ["Company Management"],
				"summary": "Retrieves all members belonging to the current user's company.",
				"description": "# Errors\nReturns an error if the user is not associated with a company or if the database query fails.",
				"operationId": "get_company_members",
				"responses": {
					"200": {
						"description": "Company members retrieved successfully",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": { "$ref": "#/components/schemas/GetCompanyMembersResponse" }
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					}
				},
				"security": [{ "bearer_auth": [] }]
			}
		},
...
	},
	"components": {
		"schemas": {
...
			"AdminUpdateMemberRequest": {
				"type": "object",
				"required": ["email", "first_name", "last_name", "role"],
				"properties": {
					"branch_id": { "type": ["string", "null"], "example": "branch-uuid-here" },
					"email": { "type": "string", "example": "user@example.com" },
					"first_name": { "type": "string", "example": "Jane" },
					"last_name": { "type": "string", "example": "Smith" },
					"role": { "type": "string", "example": "staff" }
				}
			},
...
			"ErrorResponse": {
				"type": "object",
				"required": ["error"],
				"properties": { "error": { "type": "string" } }
			},
...
			"GetCompanyMembersResponse": {
				"type": "object",
				"required": ["members"],
				"properties": {
					"members": { "type": "array", "items": { "$ref": "#/components/schemas/UserResponse" } }
				}
			},
...
			"RemoveMemberRequest": {
				"type": "object",
				"required": ["email"],
				"properties": { "email": { "type": "string", "example": "user@example.com" } }
			},
...
			"UserResponse": {
				"type": "object",
				"required": ["email", "first_name", "last_name", "role"],
				"properties": {
					"branch_id": { "type": ["string", "null"] },
					"company_name": { "type": ["string", "null"] },
					"email": { "type": "string" },
					"first_name": { "type": "string" },
					"last_name": { "type": "string" },
					"oauth_provider": { "type": ["string", "null"] },
					"role": { "$ref": "#/components/schemas/UserRole" }
				}
			},
			"UserRole": {
				"type": "string",
				"enum": ["admin", "member", "logsmart_admin", "company_manager", "branch_manager", "staff"]
			},
...
		},
		"securitySchemes": { "bearer_auth": { "type": "http", "scheme": "bearer" } }
	},
...
}